--- a/fe/fe-core/src/main/java/com/starrocks/sql/plan/PlanFragmentBuilder.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/plan/PlanFragmentBuilder.java
@@ -2422,7 +2422,7 @@ public class PlanFragmentBuilder {
                 if (constraints != null) {
                     UKFKConstraints.JoinProperty joinProperty = constraints.getJoinProperty();
                     if (joinProperty != null) {
-                        joinProperty.buildExpr(context);
+                        joinProperty.setOneMatchProbe(joinNode.getJoinOp());
                         joinNode.setUkfkProperty(joinProperty);
                     }
                 }