--- a/test/sql/test_ukfk_constraints/R/test_ukfk_constraints
+++ b/test/sql/test_ukfk_constraints/R/test_ukfk_constraints
@@ -370,3 +370,52 @@ SELECT * FROM t_uk RIGHT SEMI JOIN t_fk ON t_uk.uk = t_fk.fk AND t_uk.uk > 2 AND
 13	1	1	1	1	1	4
 14	2	2	2	2	2	4
 -- !result
+set enable_ukfk_opt=true;
+-- result:
+-- !result
+set enable_ukfk_join_reorder=false;
+-- result:
+-- !result
+SELECT * FROM t_fk JOIN t_uk ON t_uk.uk = t_fk.fk ORDER BY t_fk.id;
+-- result:
+1	1	1	1	1	1	1	1	2	3	4	5	6
+2	2	2	2	2	2	1	1	2	3	4	5	6
+3	3	3	3	3	3	1	1	2	3	4	5	6
+4	4	4	4	4	4	1	1	2	3	4	5	6
+5	5	5	5	5	5	1	1	2	3	4	5	6
+6	1	1	1	1	1	2	2	3	4	5	6	7
+7	2	2	2	2	2	2	2	3	4	5	6	7
+8	3	3	3	3	3	2	2	3	4	5	6	7
+9	4	4	4	4	4	2	2	3	4	5	6	7
+10	1	1	1	1	1	3	3	4	5	6	7	8
+11	2	2	2	2	2	3	3	4	5	6	7	8
+12	3	3	3	3	3	3	3	4	5	6	7	8
+13	1	1	1	1	1	4	4	5	6	7	8	9
+14	2	2	2	2	2	4	4	5	6	7	8	9
+15	1	1	1	1	1	5	5	6	7	8	9	0
+16	2	2	2	2	2	5	5	6	7	8	9	0
+17	3	3	3	3	3	5	5	6	7	8	9	0
+-- !result
+SELECT * FROM t_fk LEFT OUTER JOIN t_uk ON t_uk.uk = t_fk.fk ORDER BY t_fk.id;
+-- result:
+1	1	1	1	1	1	1	1	2	3	4	5	6
+2	2	2	2	2	2	1	1	2	3	4	5	6
+3	3	3	3	3	3	1	1	2	3	4	5	6
+4	4	4	4	4	4	1	1	2	3	4	5	6
+5	5	5	5	5	5	1	1	2	3	4	5	6
+6	1	1	1	1	1	2	2	3	4	5	6	7
+7	2	2	2	2	2	2	2	3	4	5	6	7
+8	3	3	3	3	3	2	2	3	4	5	6	7
+9	4	4	4	4	4	2	2	3	4	5	6	7
+10	1	1	1	1	1	3	3	4	5	6	7	8
+11	2	2	2	2	2	3	3	4	5	6	7	8
+12	3	3	3	3	3	3	3	4	5	6	7	8
+13	1	1	1	1	1	4	4	5	6	7	8	9
+14	2	2	2	2	2	4	4	5	6	7	8	9
+15	1	1	1	1	1	5	5	6	7	8	9	0
+16	2	2	2	2	2	5	5	6	7	8	9	0
+17	3	3	3	3	3	5	5	6	7	8	9	0
+18	1	1	1	1	1	None	None	None	None	None	None	None
+19	2	2	2	2	2	None	None	None	None	None	None	None
+20	3	3	3	3	3	None	None	None	None	None	None	None
+-- !result